pyproject.toml
[project]
name = "bgr-user-management"
version = "1.0.0"
description = "BGR User Onboarding - migrated from Power Automate"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "msal>=1.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]
